Avoid unnecessary rerender of RoadmapSummary
Previously, upvoting a suggestion causes the RoadmapSummary to rerender
even if the status counts didn't change. By default, useSelector() uses
strict equality to decide whether to trigger a rerender. Since we're
using an object with shape `{ status: count }`, we need to pass a second
argument to tell useSelector to use shallow equality.