Make form buttons more tab-friendly
Previously, the shift in mobile to desktop layout was implemented by
changing the flex order so that the Delete and Cancel buttons are placed
before the Submit button. Consequently, the tab order for desktop view
is reversed, which can be disorienting for users.

This commit implements the layout shift by placing a set of Delete and
Cancel buttons -- one for desktop and another for mobile -- before and
after the Submit button. This maintains the expected focus order when
pressing the tab key.